Continuing subcategories from last change
The last change was successful. I was able to add subcategories of art/sports/legos without affecting the spacing. So I continued that formatting for Week 2 and Week 3. I also corrected the text on Week 1. The last change had the text for Week 2 incorrectly under Week 1. The text treatment of "ART' etc is not ideal. I would like it to be bold and 18pt, without affecting the spacing. I will try that next.